57分页展示帖子列表【完成】

+++ models/community.go
@@ -1,7 +1,7 @@
 package models
 
 type Community struct {
-	CommunityId   int    `json:"id" db:"community_id"`
+	CommunityId   int64  `json:"id" db:"community_id"`
 	CommunityName string `json:"name" db:"community_name"`
 	Introduction  string `json:"introduction" db:"introduction"`
 }
